+++ src/client/token-provider.ts
@@ -1,0 +1,67 @@
+/**
+ * This file is part of the node-appstore-connect distribution.
+ * Copyright (c) e.GO Digital GmbH, Aachen, Germany (https://www.e-go-digital.com/)
+ *
+ * node-appstore-connect is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, version 3.
+ *
+ * node-appstore-connect is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {ClientOptions} from "./client-options";
+import * as fs from "fs";
+import {google, androidpublisher_v3} from "googleapis";
+import {JWT} from "google-auth-library";
+
+export interface AuthParams {
+    /** Google Client Email */
+    client_email: string
+    /** Google Private Key */
+    private_key: string
+}
+
+export class TokenProvider {
+
+    private client: JWT;
+    private publisher: androidpublisher_v3.Androidpublisher;
+
+    constructor(public readonly options: ClientOptions) {
+    }
+
+    private async _getClient(): Promise<JWT> {
+        if(!this.client){
+            const authData = JSON.parse(fs.readFileSync(this.options.keyPath).toString()) as AuthParams;
+            this.client = new google.auth.JWT(
+                authData.client_email,
+                undefined,
+                authData.private_key,
+                ["https://www.googleapis.com/auth/androidpublisher"]
+            );
+            await this.client.authorize();
+        }
+        return this.client;
+    }
+
+    public async getPublisher(): Promise<androidpublisher_v3.Androidpublisher> {
+        if(!this.publisher){
+            const publisher: Function = google.androidpublisher;
+            const client = await this._getClient();
+            this.publisher = publisher({
+                auth: client,
+                version: 'v3'
+            })
+        }
+        return this.publisher;
+    }
+
+    public getKeyFilePath(): string {
+        return this.options.keyPath;
+    }
+}