+++ src/release/submit-for-review-options.ts
@@ -1,0 +1,30 @@
+/**
+ * This file is part of the node-appstore-connect distribution.
+ * Copyright (c) e.GO Digital GmbH, Aachen, Germany (https://www.e-go-digital.com/)
+ *
+ * node-appstore-connect is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, version 3.
+ *
+ * node-appstore-connect is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import {ReleaseNotesInterface} from "./release-notes.interface";
+import {EditsDetailsInterface} from "../edits";
+import {TrackId} from "./track-id";
+import {LocalizedListingInterface} from "./localized-listing.interface";
+
+export interface SubmitForReviewOptions {
+    editId?: string;
+    autoAttachVersionCode?: number;
+    releaseNotes?: string | ReleaseNotesInterface | ReleaseNotesInterface[];
+    details?: EditsDetailsInterface;
+    track?: TrackId
+    listings?: LocalizedListingInterface[];
+}